requirements.txt
Django>=1.5,<1.6
django-extensions>=1.2.0
django-json-field>=0.5.5
beautifulsoup4>=4.3.1
selenium==2.35.0
